+++ packages/orq-rc/RELEASES.md
@@ -9419,3 +9419,13 @@ Based on:
 - [typescript v4.10.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.28 - packages/orq-rc
+
+## 2026-05-25 05:52:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.764.1 (2.885.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.29 - packages/orq-rc